+++ zebra-network/src/zakura/testkit/cluster.rs
@@ -207,6 +207,10 @@ mod tests {
             height,
             sapling_root: sapling::tree::NoteCommitmentTree::default().root(),
             orchard_root: orchard::tree::NoteCommitmentTree::default().root(),
+            ironwood_root: zebra_chain::ironwood::tree::NoteCommitmentTree::default().root(),
+            sapling_tx: 0,
+            orchard_tx: 0,
+            ironwood_tx: 0,
             auth_data_root: block::merkle::AuthDataRoot::from([0u8; 32]),
         }
     }